Moved pointSize in shader to double

+++ src/osgTools/PointShaderProgram.cpp
@@ -9,7 +9,7 @@ const char* vertexShaderSource = R"(
     in vec4 osg_Color;
 
     uniform mat4 osg_ModelViewProjectionMatrix;
-    uniform float pointSize;
+    uniform double pointSize;
 
     out vec4 vertexColor;
